+++ tests/gopjrt/shardy_test.go
@@ -48,7 +48,7 @@ func testShardy(t *testing.T, client *pjrt.Client) {
 	t.Run("input-data-sharding", func(t *testing.T) {
 		mesh := must1(shardy.NewDeviceMesh("data_mesh", []int{2}, []string{"data"}))
 		builder := stablehlo.New(t.Name()).WithShardy(mesh)
-		fn := builder.NewFunction("main")
+		fn := builder.Main()
 		x := must1(fn.NamedInputWithSharding("arg0", shapes.Make(dtypes.F32, 2, 3),
 			builder.NewShardingSpec().AddShardedAxis("data")))
 		reductionFn := fn.Closure()